movies_admin/requirements/base.txt
django==3.1
django-model-utils==4.0.0
psycopg2-binary==2.9.1
factory-boy==3.2.0
PyPrind==2.11.3
tqdm==4.62.1
djangorestframework==3.12.4
django-filter==2.4.0
